fix(ordering): compute AMF fill-score in i64 to avoid i32 overflow (O1)
What: feral-ordering-core's AMF (Approximate Minimum Fill) inner loop
computed its working-fill (wf) quantities in i32 and stored them in an
i32 field — the surface contribution dext*(2*deg(e)-dext-1)
(algo.rs:948/969) and the per-supervariable accumulation
wf4 + 2*nvi*wf3 (algo.rs:1020).

Why: both factors are O(n), so the products reach ~n^2 and overflow
i32 for n >= ~46k. i32::MAX is 2_147_483_647 and 46342*46341 =
2_147_534_622 already exceeds it. The wrapped value then fed the RMF
pivot score (read as f64), silently degrading ordering quality on the
large KKTs AMF exists for; in debug builds it panicked outright. MUMPS
computes the RMF in DBLE for exactly this reason.

Fix (finding's prescription, "compute wf in f64 or i64"): widened the
ws.wf field from Vec<i32> to Vec<i64>, the wf3/wf4 accumulators to i64,
and amf_bucket_of's score parameter to i64. Extracted the two cited
formulas into pure i64 helpers amf_wf_surface and amf_wf_combine (this
also dedups the surface expression and documents the i64 requirement).
The post-quantization RMF score later stored in wf is bounded by
i32::MAX-1 and cast in on store. The minimum-degree (AMD) path never
reads or writes wf and is unaffected; the Metric trait (Score=i32) is
unchanged. wf is pub but has no external readers, so the change is
fully self-contained.

Evidence: reproducing unit test amf_wf_kernels_do_not_overflow_i32
pins the kernels against the exact hand-computed i64 values of the
Amestoy 1999 formulas (ana_orderings.F:4810): amf_wf_surface(46342,
46342)=2_147_534_622, amf_wf_combine(0,46342,46342)=4_295_161_928.
RED demonstrated by temporarily narrowing both helper bodies to i32:
debug panic "attempt to multiply with overflow". GREEN with i64.
feral-ordering-core lib 57 passed (+1), invariants integration 10
passed, clippy --all-targets -D warnings clean, fmt clean. End-to-end
reproduction at n>=46k is infeasible (a genuine 46k AMF front needs
~1e9 input edges and is suppressed by dense-deferral / mass-elimination
on smaller inputs), so the deterministic kernel unit test with an
external arithmetic oracle is the reproduction. Finding O1 from
dev/research/repo-review-2026-06-09.md.

+++ crates/feral-ordering-core/src/quotient_graph/workspace.rs
@@ -85,7 +85,14 @@ pub struct Workspace {
     /// `i`, holds the running quantized RMF score; for element indices
     /// `e`, holds the lazily-cached `dext * (2*deg(e) - dext - 1)`
     /// surface contribution (sentinel `0` = "first touch this iter").
-    pub wf: Vec<i32>,
+    ///
+    /// `i64` (not `i32`): the un-quantized surface contribution has
+    /// both factors `O(n)`, so it reaches ~`n^2` and overflows `i32`
+    /// for `n` ≳ 46k before being consumed as `f64` in the RMF score
+    /// (O1, `dev/research/repo-review-2026-06-09.md`). MUMPS computes
+    /// the RMF in DBLE for the same reason. The post-quantization RMF
+    /// score stored here later is bounded by `i32::MAX - 1`.
+    pub wf: Vec<i64>,
 
     /// Generation counter for the mark array `w`.
     pub wflg: i32,
@@ -217,7 +224,7 @@ impl Workspace {
         let mut head: Vec<i32> = vec![NONE; n_buckets];
         let mut next: Vec<i32> = vec![NONE; n];
         let mut last: Vec<i32> = vec![NONE; n];
-        let mut wf: Vec<i32> = vec![0; n];
+        let mut wf: Vec<i64> = vec![0; n];
 
         let wbig = i32::MAX - n as i32;
         let wflg = 0; // clear_flag will lift to 2 on first use.
@@ -253,7 +260,7 @@ impl Workspace {
                 }
                 next[i] = inext;
                 head[deg] = i as i32;
-                wf[i] = deg as i32;
+                wf[i] = deg as i64;
             }
         }
